% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Plot Comparison of Nations in UIS Dataset}
\usage{
plot_compare(
  df,
  countries = NULL,
  top = 0,
  stat = "mean",
  graph = "bar",
  include_world_stat = TRUE,
  world_stat = "median",
  remove_regions = TRUE,
  desc = TRUE,
  axis = 0,
  round_places = 0,
  labs = TRUE,
  use_code_in_axis = TRUE,
  title = NULL,
  use_code_in_label = TRUE,
  color_palette = NULL
)
}
\arguments{
\item{df}{Dataframe.
The \code{data.frame} to be filtered.
This should be created using the \code{uis_filter_indicator} function.}

\item{countries}{Character Vector.
The three-letter codes of countries to compare.
The function will highlight and label these countries.}

\item{top}{Numerical vector or numerical variable.
This will filter the top countries.
Negative numbers will filter the bottom countries.
Defaults to all countries.
See examples for more details.}

\item{stat}{\code{'mean'}, \code{'median'}, \code{'sd'}, \code{'min'}, and \code{'max'}.
Since the output of \code{uis_filter_indicator} contains multiple years,
this is the statistic that will be calculated
per country for all the available years.}

\item{graph}{\code{'bar'} or \code{'segment'}.
Type of graph to be plotted. Defaults to \code{'bar'}.}

\item{include_world_stat}{Logical. Defaults to \code{TRUE}.
If \code{FALSE}, it will exclude the world statistic to be highlighted and labeled.}

\item{world_stat}{\code{'mean'} or \code{'median'}.
The world statistics to calculate. Defaults to \code{'median'}.
Will be ignored if \code{include_world_stat} is \code{FALSE}.}

\item{remove_regions}{Logical.
This will remove the regional summaries in the UIS dataset.
Defaults to \code{TRUE}.}

\item{desc}{Logical. Sorts the plot into descending order. Defaults to \code{TRUE}.}

\item{axis}{\code{0} or \code{1}. Orientation of plot.}

\item{round_places}{Integer. The number of places to round in calculations.
Defaults to 0.}

\item{labs}{Logical. Determines whether to show the labels in the plot.
Determines}

\item{use_code_in_axis}{Determines whether to show the
code on the axis instead of the whole name of the country. Defaults to TRUE.}

\item{title}{Character. The title to show on top.}

\item{use_code_in_label}{Determines whether to show the
code on the labels instead of the whole name of the country. Defaults to TRUE.}

\item{color_palette}{RColorBrewer palettes. EXPERIMENTAL.
Sets the color of the highlighted bars/segments.}
}
\description{
\code{plot_compare} plots a ggplot comparing the statistics of
nations with the dataset made using the function \code{uis_filter_indicator}
}
\examples{
df_source <- read.csv("NATMON_DS_04012021100059822.csv")
df <- uis_filter_indicator(df_source,"FOSGP_5T8_F900")

plot_compare(df)
plot_compare(df1,
             countries = c("PHL"),
             round_places = 3) +
   theme(axis.text.y = element_blank())
plot_compare(df,
           countries = c("PHL","MMR","ITA"),
           stat = "median",
           graph = "segment",
           top = 30,
           axis=1,
           round_places = 3)
plot_compare(df,
           countries = c("PHL","MMR","ITA"),
           stat = "median",
           world_stat = "mean",
           top = seq(20,40),
           axis=1,
           round_places = 3)
plot_compare(df,
           countries = c("PHL","MMR","ITA"),
           stat = "median",
           graph = "segment",
           axis=1,
           round_places = 3,
           top = -20
           include_world_stat = FALSE,
           use_code_in_label = FALSE,
           color_palette = "BuGn")}
\seealso{
}
