% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_indicator_ids.R
\name{show_indicator_ids}
\alias{show_indicator_ids}
\title{Show UIS Indicators}
\usage{
show_indicator_ids(df)
}
\arguments{
\item{df}{The \code{data.frame} to be filtered. Leave the UIS dataset as is}
}
\description{
\code{show_indicator_ids} shows the indicators and their corresponding indicator IDs on the current dataframe
}
