% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis_filter_country.R, R/uis_filter_indicator.R
\name{uis_filter_country}
\alias{uis_filter_country}
\alias{uis_filter_indicator}
\title{Filter UNESCO Institute of Statistics Dataset}
\usage{
uis_filter_country(df, country_code, years = NULL, indicator_ids = NULL)

uis_filter_indicator(df, indicator_id, years = NULL, countries = NULL)
}
\arguments{
\item{df}{The \code{data.frame} to be filtered. Leave the UIS dataset as is for it should have the Indicator ID as its first column, regardless of its column name, and it should contain the following columns:
\itemize{
\item \code{Indicator} : Indicator names;
\item \code{LOCATION}: Three-letter ISO country codes;
\item \code{Country} : Official country names;
\item \code{TIME} : Year; and
\item \code{Value} : Numerical values.
}}

\item{country_code}{The three-letter ISO code of the country to be filtered. It should be a character object with length 1.}

\item{years}{The years to filter. It should be a numerical vector. Defaults to all available years in the dataset.}

\item{indicator_ids}{The IDs of the indicator to be filtered as character vector. Defaults to all indicators in the asked  country.}

\item{indicator_id}{The ID of the indicator to be filtered.
Try \code{df[match(unique(df[,1]),df[,1]),c(1,2)]} to know
more about the indicator IDs.}

\item{countries}{The countries to be filtered. It should be a character vector in the country code format.
Defaults to all countries (including the UIS Regional summaries, if available).}
}
\description{
\code{uis_filter_country} filters UIS Dataset using a single country code.

\code{uis_filter_indicator} filters UIS dataset using a single Indicator ID.

This dataset must be downloaded from the
\href{http://data.uis.unesco.org/}{UNESCO Institute of Statistics (UIS) Data Browser}
as a \code{.csv} file in default (,) format.
}
\examples{
df <- read.csv("NATMON_DS_04012021100059822.csv")

uis_filter_country(df,
                   "PHL")
uis_filter_country(df,
                   "PHL",
                    indicator_ids=c("FOSGP_5T8_F900","FOSGP_5T8_F300","FOSGP_5T8_FUK"))
uis_filter_country(df,
                   "PHL",
                   years=seq(2010,2013),
                   indicator_ids=c("FOSGP_5T8_F900","FOSGP_5T8_F300","FOSGP_5T8_FUK"))
uis_filter_indicator(df,
                     "FOSGP_5T8_F500")
uis_filter_indicator(df,
                     "FOSGP_5T8_F500",
                      countries=c("PHL","USA","CHN"))
uis_filter_indicator(df,
                      "FOSGP_5T8_F500",
                       years=seq(2010,2013),
                       countries=c("PHL","USA","CHN"))
}
